AI system functions as designed

+++ EntityAntFarm/Generators.cpp
@@ -4,28 +4,34 @@ Entity TrailGenerator::new_trail(int32_t trail, std::array<int32_t, 3> position)
 {
     Entity newEntity = this->entityManager->create_entity();
 	this->trailManager->add_entity_component({ newEntity, trail });
-	this->positionManager->add_entity_component({ newEntity, position });
+	this->positionManager->add_position_component({ newEntity, position });
 	this->colorManager->add_entity_component({ newEntity, color_to_int({ 0, 0, 0, 0 }) });
+	this->createdTrails.insert({ newEntity, true });
     return newEntity;
 }
 
 void TrailGenerator::destroy_trail(Entity e) {
-	this->trailManager->remove_entity_component(e);
-	this->positionManager->remove_entity_component(e);
-	this->colorManager->remove_entity_component(e);
-	this->entityManager->destroy_entity(e);
+	auto it = this->createdTrails.find(e);
+	if (it != this->createdTrails.end()) {
+		this->trailManager->remove_entity_component(e);
+		this->positionManager->remove_position_component(e);
+		this->colorManager->remove_entity_component(e);
+		this->entityManager->destroy_entity(e);
+		this->createdTrails.erase(it);
+	}
 }
 
 Entity AntGenerator::new_ant(std::array<int32_t, 3> position)
 {
 	Entity newEntity = this->entityManager->create_entity();
-	this->positionManager->add_entity_component({ newEntity, position });
+	this->positionManager->add_position_component({ newEntity, position });
 	this->colorManager->add_entity_component({ newEntity, color_to_int({ 255, 255, 255, 255 }) });
 	this->velocityManager->add_entity_component({ newEntity, {0, 0, 0, 0} });
-	this->foodManager->add_entity_component({ newEntity, 0 });
+	this->foodManager->add_entity_component({ newEntity, 0, TRANSPORTER });
 	this->aiManager->add_entity_component({ newEntity, AICodes::Seeking });
 	this->historyManager->add_entity_component({ newEntity, 0 });
 	this->collisionManager->add_entity_component({ newEntity, {}, {} });
+	this->createdAnts.insert({ newEntity, true });
 	return newEntity;
 }
 
@@ -33,9 +39,10 @@ Entity FoodGenerator::new_food(int32_t food, std::array<int32_t, 3> position)
 {
 	Entity newEntity = this->entityManager->create_entity();
 	
-	this->foodManager->add_entity_component({ newEntity, food });
+	this->foodManager->add_entity_component({ newEntity, food, GIVER });
 	this->colorManager->add_entity_component({ newEntity, color_to_int({255, 150, 0, 255}) });
-	this->positionManager->add_entity_component({ newEntity, position });
+	this->positionManager->add_position_component({ newEntity, position });
+	this->createdFood.insert({ newEntity, true });
 
 	return newEntity;
 }
@@ -54,8 +61,12 @@ void FoodGenerator::new_food_cluster(int32_t n, uint32_t food, std::array<int32_
 
 void FoodGenerator::destroy_food(Entity e)
 {
-	this->entityManager->destroy_entity(e);
-	this->positionManager->remove_entity_component(e);
-	this->colorManager->remove_entity_component(e);
-	this->foodManager->remove_entity_component(e);
+	auto it = this->createdFood.find(e);
+	if (it != this->createdFood.end()) {
+		this->entityManager->destroy_entity(e);
+		this->positionManager->remove_position_component(e);
+		this->colorManager->remove_entity_component(e);
+		this->foodManager->remove_entity_component(e);
+		this->createdFood.erase(it);
+	}
 }
